faz validação de erros funciona

+++ src/main/java/com/ativ1/springboot/controllers/CustomerController.java
@@ -2,8 +2,13 @@ package com.ativ1.springboot.controllers;
 
 import com.ativ1.springboot.domain.Customer;
 import com.ativ1.springboot.repository.CustomerRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
 import java.util.List;
+import java.util.Map;
 
 //ENDPOINT DE CLIENTES
 @RestController
@@ -16,8 +21,9 @@ public class CustomerController {
     }
 
     @PostMapping
-    public Customer Create(@RequestBody Customer customer) {
-        return repository.save(customer);
+    public ResponseEntity<Customer> create(@RequestBody Customer customer) {
+        var customerCreated = repository.save(customer);
+        return ResponseEntity.status(201).body(customerCreated);
     }
 
     @GetMapping
@@ -26,8 +32,10 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}")
-    public Customer ListByID(@PathVariable Long id) {
-        return repository.findById(id).orElse(null);
+    public ResponseEntity<?> ListByID(@PathVariable Long id) {
+        var customer =  repository.findById(id);
+        if(customer.isPresent()) return ResponseEntity.ok(customer.get());
+        return ResponseEntity.status(404).body(Map.of("error", "Usuário não encontrado."));
     }
 
     @PutMapping("/{id}")
@@ -38,11 +46,19 @@ public class CustomerController {
             customer.setAge(newCustomer.getAge());
             customer.setGender(newCustomer.getGender());
             return repository.save(customer);
-        }).orElse(null);
+        }).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
     }
 
     @DeleteMapping("/{id}")
-    public void Delete(@PathVariable Long id) {
-        repository.deleteById(id);
+    public ResponseEntity<?> Delete(@PathVariable Long id) {
+        var customer = repository.findById(id);
+        if (customer.isPresent()) {
+            repository.deleteById(id);
+            return ResponseEntity.status(200)
+                    .body(Map.of("message", "Usuário deletado com sucesso."));
+        }
+        return ResponseEntity
+                .status(404)
+                .body(Map.of("error", "Usuário não encontrado."));
     }
 }
